chore: update derive_more to v2

+++ websockets/chat-broker/src/message.rs
@@ -5,12 +5,12 @@ use actix::prelude::*;
 pub struct ChatMessage(pub String);
 
 #[derive(Clone, Message)]
-#[rtype(result = "usize")]
+#[rtype(result = "u64")]
 pub struct JoinRoom(pub String, pub Option<String>, pub Recipient<ChatMessage>);
 
 #[derive(Clone, Message)]
 #[rtype(result = "()")]
-pub struct LeaveRoom(pub String, pub usize);
+pub struct LeaveRoom(pub String, pub u64);
 
 #[derive(Clone, Message)]
 #[rtype(result = "Vec<String>")]
@@ -18,4 +18,4 @@ pub struct ListRooms;
 
 #[derive(Clone, Message)]
 #[rtype(result = "()")]
-pub struct SendMessage(pub String, pub usize, pub String);
+pub struct SendMessage(pub String, pub u64, pub String);
